requirements.txt
asgiref==3.11.1
Django==6.0.2
django-allauth==65.14.3
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
